# `<filesystem>`: Improve `symlink_status` performance