--- a/benchmarks/src/filesystem.cpp
+++ b/benchmarks/src/filesystem.cpp
@@ -11,6 +11,7 @@ void symlink_status(benchmark::State& state) {
         std::error_code ec;
         const auto status = std::filesystem::symlink_status(path, ec);
         benchmark::DoNotOptimize(status);
+        benchmark::DoNotOptimize(ec);
     }
 }