@StephanTLavavej 2024-11-11
Thanks! On my system (5950X, 24H2, Samsung SSD 980 PRO), I observe much less of a performance improvement, but still major:  21126 ns => 10098 ns, for a speedup of 2.1x.