@StephanTLavavej 2024-11-14
I'm mirroring this to the MSVC-internal repo - please notify me if any further changes are pushed.