@StephanTLavavej 2024-11-14
Thanks again, and congratulations on your first microsoft/STL commit! :heart_eyes_cat: :tada: :floppy_disk:

This change is expected to ship in VS 2022 17.13 Preview 3.